Embed preguide packages in preguide itself (#17)
Theoretically it is possible to write a CUE package for a guide without
importing the preguide CUE packages: the only validation that ultimately
matters is when preguide itself runs. Hence it's possible that a user
(guide author) might not actually have the preguide packages available
via their cue.mod (they might only be depending on the preguide Go main
package).

In this case we cannot rely on being able to load the preguide packages
for validation purposes at runtime.

Hence we embed those two packages in preguide itself.

Note the tests themselves are written with guides that _do_ import the
preguide packages, hence we need to continue to vendor preguide itself
as part of each testscript script. Added a comment to this effect.