Remove Script from prestep output
For now presteps will only emit environment variables. Therefore we
remove the Script field, but leave it such that presteps return a struct
for future extensibility.

We also change the Vars field to a []string of environment values (such
that can be added to os/exec.Cmd.Env).

We also switch to using the go1.14.6 image which pre-initialises the git
config (this was previously done in a prestep Script in the git sanitise
test).