Add support for multiple scenarios (#52)
Scenarios are variations on a guide just as language translations are.
At the finest level of granularity, steps can be defined per language
per scenario. At the highest level they can be defined per language.

In a later iteration of this work, we will support either (at the moment
we only support per language, which implies for all scenarios).