Escape {{X}} delimited blocks in markdown output
Otherwise Liquid (part of Jekyll) tries to interpret them. This also
required a bit of refactoring to include the variables that resulted
from a prestep in the out CUE file. Because in the case the inputs to a
guide have not changed, we still need to perform this output
normalisation and escaping, and that requires that we know what were
(and are) the variable names returned by the presteps.