cmd/preguide: add -runargs flag
-runargs allows the caller to pass arbitrary flags to docker when the
guide is "run", i.e. when the bash script that represents the file is
run. These are specified per terminal (despite us only supporting one
terminal for now) in order that different flags can be passed to each
(there are scenarios where unconditionally passing the same flags to all
terminals would be wrong, so we leave this to the caller).

Note these flags are in no way validated; any errors will be reported
back by the docker process and "bulbbled up".

Fixes #78