cmd/preguide: accept args to gen (#88)
The gen command now works in one of two ways. With no args and no -dir
flag, we behave as if -dir . has been passed. When -dir is passed, no
args can be passed. When args are passed, no -dir flag can be specified.

When operating in -dir mode, we also generate a guide structures CUE
file in the directory passed to -dir.