cmd/preguide: when we docker run self use the same convention (#101)
When we run self in dev mode (i.e. tests) we don't have debug.BuildInfo.
Hence we can't follow the "prod" code path of using a docker hub image.
This is fine, but we need to use the same convention of zero entrypoint
and passing all arguments as a command either way.

Fix that.

Also switch to using buildkit for building self, and add a step of
checking that docker build self to the main CI workflow, leaving the
"release" to happen on the post-merge "Docker self" workflow.