cmd/preguide: do not pass USER_UID and USER_GID to the running script (#102)
Not necessary. And interacts badly when run from different platforms.

Instead, ensure that we create a scripts file for the bind mount that
has more liberal permissions, and hence is not susceptible to the
UID/GID of the user within the script container being different to that
of the host.

Also introduce PREGUIDE_DEVEL_IMAGE env var in order to override the
logic on use self or a published Dockage image for running self in a
container. This is useful when we are locally testing a new version of
preguide (i.e. a remote version exists) in, say, play-with-go, _without_
the images for that version having been published.