cmd/preguide: support marking a command block as random output (#116)
Command blocks that are marked in this way have their output included
along with the result of prestep variables in the first step of
sanitisation of output. Then the existing command sanitisers run.

As part of this change we delete the git sanitiser, because that output
can more easily be handled by using the command block sanitisation.