cmd/preguide: remove concept of trimmed output from out package
The use/need for defining definitions within the out package remains to
be really proven.

Stable output like go version can simply be shown as a command block,
indeed that's more useful to the user

Unstable output cannot be included in prose in any case because it would
only be the sanitised output, not the actual output.

Rather than totally removing this concept, we instead remove the
TrimmedOutput field to simplify it.