cmd/preguide: log the bash script that ran on output parse error
We have seen one or two random failures where the output of the bash
script could not be parsed because some random output appears after an
exit code but before the next output fence.

Output the bash script that ran in that case to try and understand what
is going on.