cmd/preguide: sanity check guides with a separate load (#123)
cuelang.org/issue/567 has the unfortunate side effect of rendering our
Unify check useless. Therefore, until that issue is fixed, we do a
sanity check load using a temporary CUE file that imports the package we
are trying to load, along with preguide, and unifies the two instances.
This means that if the guide instance does not satisfy #Guide then the
loaded value will be in error.