cmd/preguide: do not write presteps and terminals to a guide log
Whilst this important detail for a guide, it is included in the
generated out package for a guide. The log is meant to be human
readable. And it is meant to be a quick way of determining how the
output of a guide has changed. Therefore, to include version information
in these files makes the diff too noisy; you always end up ignorning the
part of the diff that relates to version information, which means that
the signal to noise ratio is too low. Hence these diffs start to be
ignored.