cmd/preguide: support sanitisers for comparison output (#144)
go get is unfortunately not deterministic with its output. The canonical
example is:

	$ go get golang.org/x/tools@v0.0.0-20201105220310-78b158585360
	go: downloading golang.org/x/tools v0.0.0-20201105220310-78b158585360
	go: found golang.org/x/tools/cmd/stringer in golang.org/x/tools v0.0.0-20201105220310-78b158585360
	go: downloading golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1
	go: downloading golang.org/x/mod v0.3.0

The order of the deps golang.org/x/xerrors and golang.org/x/mod being
downloaded can, and does, vary, even with GOMAXPROCS=1 set (because the
ordering and concurrency of the download output is not controlled via
GOMAXPROCS).

This often gets triggered in CI, when PREGUIDE_SKIP_CACHE=true.
Therefore, when running in this mode, if we would otherwise have had a
cache hit (i.e. inputs are hash equal), we need to compare the generated
output of the guide on disk and the re-generated output (because
PREGUIDE_SKIP_CACHE=true), but do so post sanitisation of what we are
now calling ComparisonOutput. If the comparison output is identical,
then we do _not_ write the re-generated guide to disk.