cmd/preguide: fix escaping of braces (#178)
This commit also fixes the escaping of braces because the current
approach is a) wrong, b) poorly documented and c) interpreted
incorrectly when it comes to what is possible from the author's
perspective.

There are broadly speaking two cases to consider:

1. when we are outputting markdown to a file that will then be
   interpreted by a processor that interprets template blocks of some
   sort. e.g. Jekyll. In which case we need to consider when these
   template blocks are interpretted as well as regular markdown rules.
2. when we are not outputting to such a system, in which case only
   regular markdown rules apply.

Regarding category 1, Jekyll interprets {{,}} and {%,%} pairs wherever
it seems them _before_ markdown rules are applied.

Hence, the first thing to do is determine if and when it should be
possible for the user to specify templates/blocks for case 1.

The conclusion we have drawn for preguide is that it should be possible
to specify these templates in prose, but not in commands/code. Reason
being, it is impossible for the expansion of templates to happen before
commands are executed/code uploaded. But also that the author has not
control over the output of the command steps. Hence we can only assume
in those cases that { and } characters are meant literally.

In prose, if the user wants to render a literal { or }, then it's easy
to use the HTML entities &#123; and &#125;. If they want to include a
Jekyll directive they can using {{ .blah }} - preguide will leave this
alone (but the GitHub rendering of this will likely be wrong, hence
probably not a good idea to include directives).

In commands/code, { and } are always literal.

So to the escaping rules for { and }. The user takes care of the
escaping in prose. In commands/code, it's simply a case of replacing {
and } with their HTML entity equivalents (because commands and code are
now rendered as <pre><code> blocks).

Note that JavaScript processing of {{{.ENV}}} variable templates still
works with HTML entities, because the JavaScript evaluates over the
resulting unicode graphemes.

So, TL;DR:

* if the authors wants to include { or } in their prose, they should use
  &#123; and &#125;
* { and } appearing in command/code blocks will be treated literally
* It is possible to include {{ }} Jekyll directives in prose, assuming
  that the author has configured the guide to use different delimiters
* {{{ }}} script directives can only reference steps and refs (with a .
  shorthand for refs). It is not possible to reference {{{ .ENV }}}
  results from the prestep as it once was
* Script output (commands and code) will be sanitised for random values
  and prestep/environment values. Environment values will be templated
  using the script delimiters, generally {{{, }}}