all: refactor ahead of multi language and scenario (#189)
This includes dropping support for out package Defs. This is not even
used in PWG, is vaguely defined, and poorly implemented. Remove this
(again) for now. If we want to re-add this in a multi scenario/langauge
world we can, but the Defs themselves in an out package will also need
to be at the granularity of language + scenario. Hence it will require
some more thinking through.