cmd/preguide: ignore build constraints when loading config (#190)
This is a hack. But it works for now. To properly fix this hack we need
to provide flags for specifying the tags that should be passed during
the loading of config _and_ the guides themselves (because they might be
different... and CUE is stricter about tags being used in files)