cmd/preguide: sanitise go version output
go version contain arch-specific information. This therefore needs to be
sanitised for the benefit of comparison and stable output.

As a later exercise we might make this a value that is template-ised, so
that the real value specific to the user's system is inserted at the
point of reading a guide.