cmd/preguide: better support for GitHub markdown output
We previously fell back to <pre><code> blocks for code file output in
GitHub markdown mode in order that we could clearly show diffs using <b>
elements. However, this suffers from the very obvious disadvantage that
code is not syntax highlighted. For some (most?) guides, this will
result in large blocks of non-highlighted code which is hard to read for
many (most?).

Revert therefore to using ``` code blocks which do enable syntax
highlighting for now.

We can revisit approaches for better highlighting diffs with respect to
other code blocks in the future.